move input_metadata
