+++ python/sglang/srt/model_executor/cuda_graph_runner.py
@@ -29,12 +29,12 @@ from sglang.srt.layers.logits_processor import (
     LogitsMetadata,
     LogitsProcessor,
 )
-from sglang.srt.managers.schedule_batch import (
+from sglang.srt.managers.schedule_batch import ScheduleBatch
+from sglang.srt.model_executor.forward_batch_info import (
+    ForwardMode,
     InputMetadata,
-    ScheduleBatch,
     init_flashinfer_args,
 )
-from sglang.srt.model_executor.forward_batch_info import ForwardMode
 from sglang.srt.utils import monkey_patch_vllm_all_gather
